+++ src/Resources/app/storefront/src/plugin/listing/elio-listing-extension.plugin.js
@@ -8,10 +8,18 @@ export default class ElioListingExtensionPlugin extends window.PluginBaseClass {
         listingPropertiesSelector: '.e-elio-data-discovery-listing-properties',
         filterPanelSelector: '.filter-panel-items-container',
         filterPanelActiveSelector: '.filter-panel-active-container',
-        filterPanelItemDropdownSelector: '.filter-panel-item-dropdown'
+        filterPanelItemDropdownSelector: '.filter-panel-item-dropdown',
+        active: window.elioDataDiscovery.global.active === "1" && (
+            (window.elioDataDiscovery.global.searchUseElioDataDiscovery === "1" && document.body.classList.contains('is-act-search')) ||
+            (window.elioDataDiscovery.global.listingUseElioDataDiscovery === "1" && document.body.classList.contains('is-ctl-navigation'))
+        )
     });
 
     init () {
+        if (!this.options.active) {
+            return;
+        }
+
         const parentFilterPanelElement = DomAccess.querySelector(document, this.options.parentFilterPanelSelector)
         this.listing = window.PluginManager.getPluginInstanceFromElement(
             parentFilterPanelElement,