+++ src/Storefront/Controller/SuggestController.php
@@ -38,8 +38,7 @@ use Elio\ElioDataDiscovery\Api\Search\Response\SuggestionResponse;
 use Elio\ElioDataDiscovery\Api\Search\SuggestApi;
 use Elio\ElioDataDiscovery\Configuration\ElioDataDiscoveryConfigServiceInterface;
 use Elio\ElioDataDiscovery\Core\Logging\ElioDataDiscoveryLogTrait;
-use Elio\ElioDataDiscovery\Core\Sync\DataTypes\ProductDataType;
-use Elio\ElioDataDiscovery\Core\Util\StripClassPathUtil;
+use Elio\ElioDataDiscovery\Core\Suggest\SuggestRequestBuilder;
 use Psr\Log\LoggerInterface;
 use Shopware\Core\Content\Product\SalesChannel\Search\AbstractProductSearchRoute;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
@@ -68,6 +67,7 @@ class SuggestController extends SearchController
     /**
      * @param ElioDataDiscoveryConfigServiceInterface $configService
      * @param SuggestApi $suggestApi
+     * @param SuggestRequestBuilder $suggestRequestBuilder
      * @param SearchPageLoader $searchPageLoader
      * @param SuggestPageLoader $suggestPageLoader
      * @param AbstractProductSearchRoute $productSearchRoute
@@ -76,6 +76,7 @@ class SuggestController extends SearchController
     public function __construct(
         private readonly ElioDataDiscoveryConfigServiceInterface $configService,
         private readonly SuggestApi $suggestApi,
+        private readonly SuggestRequestBuilder $suggestRequestBuilder,
         SearchPageLoader $searchPageLoader,
         SuggestPageLoader $suggestPageLoader,
         AbstractProductSearchRoute $productSearchRoute,
@@ -107,11 +108,7 @@ class SuggestController extends SearchController
         $searchTerm = $request->get('search') ?? '*';
 
         try {
-            $suggestRequest = new SuggestRequest('');
-            $suggestRequest->setQuery($searchTerm);
-            if ($config->isSuggestToggleProductType()) {
-                $suggestRequest->setType(StripClassPathUtil::stripClassPath(ProductDataType::class));
-            }
+            $suggestRequest = $this->suggestRequestBuilder->build($searchTerm, new SuggestRequest(''), $config, $context);
             $resultCollection = $this->suggestApi->suggest($suggestRequest, $context);
 
             /** @var SuggestionResponse|null $suggestionResponse */